% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{ler_pnad}
\alias{ler_pnad}
\title{Title}
\usage{
ler_pnad(ano, trimestre)
}
\arguments{
\item{ano}{Ano}

\item{trimestre}{Trimestre}
}
\value{
Uma tibble
}
\description{
Title
}
\examples{
\dontrun{res <- ler_pnad(2018, 3)}
}
