% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.R
\name{ler_pnad_anual}
\alias{ler_pnad_anual}
\title{Ler dados da PNADc Anual}
\usage{
ler_pnad_anual(ano)
}
\arguments{
\item{ano}{Ano da pesquisa}
}
\value{
O Data frame com os microdados
}
\description{
Ler dados da PNADc Anual
}
\examples{
ler_pnad_anual(2022)
}
