% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participacao.R
\name{participacao}
\alias{participacao}
\title{Participação}
\usage{
participacao(df, variavel)
}
\arguments{
\item{df}{df da PNAD resultado de ler_pnad}

\item{variavel}{Variável que será contada}
}
\value{
Um data.frame com propoção
}
\description{
Participação
}
